Some refactoring + unit tests
