+++ FromSoftwarefileManager.Test/DatabaseEngineProviderTests.cs
@@ -1,57 +1,89 @@
-﻿using System.IO;
+﻿using System;
+using System.Data.Entity;
+using System.IO;
+using System.Linq;
+using System.Reflection;
 using System.Threading.Tasks;
 using FromSoftwareStorage;
+using FromSoftwareStorage.EntityModel;
 using NUnit.Framework;
+using Database = FromSoftwareStorage.Database;
 
 namespace FromSoftwareGameSaves.Test
 {
     [TestFixture]
     public class DatabaseEngineProviderTests
     {
-        private const string DefaultPassword = "123456789";
-        private DatabaseEngineProvider _engine;
+        private const string ConnectionStringName = "DataEntities";
+        private DatabaseEngineInstaller _installer;
+        private DatabaseEngineProvider _provider;
 
         [SetUp]
-        public void Setup()
+        public async Task Setup()
         {
-            _engine = new DatabaseEngineProvider();
+            _installer = new DatabaseEngineInstaller();
+            _provider = new DatabaseEngineProvider();
 
-            if (_engine.DataBaseExists)
-                File.Delete(_engine.DatabaseFileName);
+            Uninstall();
+
+            await _installer.InstallDatabaseAsync();
+        }
+
+        private void Uninstall()
+        {
+            if (_provider.HasDatabasePassword)
+            {
+                File.Delete(_installer.RsaPrivateKeyFileName);
+                File.Delete(_installer.RsaPublicKeyFileName);
+            }
+
+            if (_provider.IsDatabaseInstalled)
+            {
+                File.Delete(_installer.DatabaseFileName);
+            }
         }
 
         [Test]
-        public void TearDown()
+        public async Task InsertNewGameTest()
         {
-            if (_engine.HasPrivateKey)
-                File.Delete(_engine.RsaPrivateKeyFileName);
+            IDatabaseProvider databaseProvider = Database.DatabaseProvider;
+            using (DataEntities dataEntities = databaseProvider.GetEntities(ConnectionStringName))
+            {
+                DbSet<Game> dataEntitiesGames = dataEntities.Games;
 
-            if (_engine.HasPublicKey)
-                File.Delete(_engine.RsaPublicKeyFileName);
+                Game newGame = dataEntitiesGames.Create();
+                newGame.Name = "Test";
+                newGame.ChangeDate = DateTime.Now;
+                newGame.DefaultFileName = "Test.txt";
+                newGame.Directory = "Test";
+                newGame.FileSearchPattern = "*.txt";
+                newGame.ReadOnly = false;
+                newGame.Folder = new Folder() { Name = "Test", FolderPath = @"C:\", ReadOnly = false };
+                newGame.Image = new Image() { ImageFile = await GetBinaryResourceAsync("FromSoftwareGameSaves.Test.Test.Test.jpg") };
+                dataEntitiesGames.Add(newGame);
+                dataEntities.SaveChanges();
+
+                IQueryable<Game> query = from game in dataEntitiesGames
+                    where game.Name.Equals("Test")
+                    select game;
+                var gameWihTestName = query.FirstOrDefault();
+
+                Assert.That(gameWihTestName, Is.EqualTo(newGame));
+            }
+   
         }
 
-        [Test]
-        public void CreateDataWithoutPasswordTest()
+        public static async Task<byte[]> GetBinaryResourceAsync(string resourceName)
         {
-            _engine.CreateDatabase();
-            Assert.That(File.Exists(_engine.DatabaseFileName), Is.True);
-        }
+            Assembly assembly = typeof(DatabaseEngineProviderTests).Assembly;
+            using (Stream resourceStream = assembly.GetManifestResourceStream(resourceName))
+            {
+                if (resourceStream == null) return new byte[0];
+                byte[] bufferBytes = new byte[resourceStream.Length];
+                await resourceStream.ReadAsync(bufferBytes, 0, bufferBytes.Length);
 
-        [Test]
-        public void CreateDataWithPasswordTest()
-        {
-            _engine.CreateDatabase(DefaultPassword);
-            Assert.That(File.Exists(_engine.DatabaseFileName), Is.True);
-            Assert.That(File.Exists(_engine.RsaPublicKeyFileName), Is.True);
-            Assert.That(File.Exists(_engine.RsaPrivateKeyFileName), Is.True);
-        }
-
-        [Test]
-        public async Task InstallDatabaseWithoutPasswordTest()
-        {
-           bool installSucceeded = await _engine.InstallDatabaseAsync();
-           Assert.That(installSucceeded, Is.True);
-           Assert.That(File.Exists(_engine.DatabaseFileName), Is.True);
+                return bufferBytes;
+            }
         }
     }
 }
