+++ FromSoftwarefileManager.Test/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="EntityFramework" version="5.0.0" targetFramework="net472" />
   <package id="NUnit" version="3.11.0" targetFramework="net461" />
 </packages>
