+++ FromSofwareStorage/Database.cs
@@ -1,0 +1,12 @@
+﻿namespace FromSoftwareStorage
+{
+    public static class Database
+    {
+        public static IDatabaseProvider DatabaseProvider { get; }
+
+        static Database()
+        {
+            DatabaseProvider = new DatabaseEngineProvider();
+        }
+    }
+}
